Move the HTML template code into formvalidator.go as it is more closely related to that functionality. memberservice.go is now pretty much restricted to the main function.
