+++ memberservice.go
@@ -37,18 +37,8 @@ import (
 	"html/template"
 	"log"
 	"net/http"
-	"net/url"
 	"os"
 )
-
-var fmap = template.FuncMap{
-	"html": template.HTMLEscaper,
-	"url":  UserInputFormatter,
-}
-
-func UserInputFormatter(v ...interface{}) string {
-	return template.HTMLEscapeString(url.QueryEscape(v[0].(string)))
-}
 
 func main() {
 	var help bool
